working IP mapping with apis

+++ scripts/deploy.sh
@@ -10,6 +10,8 @@ default_helper() {
 
         Args:
             --target <local|EKS|Azure|GCP> - specify the deployment target for which cluster (default is local)
+            --no-deploy: Don't deploy the helm chart, only create the helm chart
+            --no-helmify: Don't create the helm chart
     "
 }
 
@@ -20,6 +22,9 @@ helmify(){
     port=$(yq e '.project.port' $CURRENT_DIR/$NAME/scaffold.yaml)
     docker_repo=$(yq e '.project.docker-repo' $CURRENT_DIR/$NAME/scaffold.yaml)
     type=$(yq e '.project.type' $CURRENT_DIR/$NAME/scaffold.yaml)
+
+    env_vars=$(yq e '.resources[].env-remote' $CURRENT_DIR/manifest.yaml)
+    IFS=$'\n' read -r -d '' -a env_vars <<< "$env_vars"
 
     if [ "$docker_run" == "null" ]; then
         print_warning "Docker Image is not built for $NAME component. Please build and push the image"
@@ -35,6 +40,12 @@ helmify(){
             -e "s#{{TAG}}#latest#" \
             -e "s#{{SERVICE_TYPE}}#None#" \
             "$KUBEFS_CONFIG/scripts/templates/deployment/helm-values.conf" > "$CURRENT_DIR/$NAME/deploy/values.yaml"
+        
+        for env in "${env_vars[@]}"; do
+            IFS='=' read -r -a env_parts <<< "$env"
+            yq e ".env += [{\"name\" : \"${env_parts[0]}\", \"value\": \"${env_parts[1]}\"}]" $CURRENT_DIR/$NAME/deploy/values.yaml -i
+        done
+        
     }
 
     helmify_frontend(){
@@ -46,6 +57,11 @@ helmify(){
             -e "s#{{TAG}}#latest#" \
             -e "s#{{SERVICE_TYPE}}#LoadBalancer#" \
             "$KUBEFS_CONFIG/scripts/templates/deployment/helm-values.conf" > "$CURRENT_DIR/$NAME/deploy/values.yaml"
+       
+        for env in "${env_vars[@]}"; do
+            IFS='=' read -r -a env_parts <<< "$env"
+            yq e ".env += [{\"name\" : \"${env_parts[0]}\", \"value\": \"${env_parts[1]}\"}]" $CURRENT_DIR/$NAME/deploy/values.yaml -i
+        done
     }
 
     helmify_api(){
@@ -57,6 +73,11 @@ helmify(){
             -e "s#{{TAG}}#latest#" \
             -e "s#{{SERVICE_TYPE}}#ClusterIP#" \
             "$KUBEFS_CONFIG/scripts/templates/deployment/helm-values.conf" > "$CURRENT_DIR/$NAME/deploy/values.yaml"
+        
+        for env in "${env_vars[@]}"; do
+            IFS='=' read -r -a env_parts <<< "$env"
+            yq e ".env += [{\"name\" : \"${env_parts[0]}\", \"value\": \"${env_parts[1]}\"}]" $CURRENT_DIR/$NAME/deploy/values.yaml -i
+        done
     }
     
     case "$type" in
@@ -72,6 +93,8 @@ parse_optional_params(){
     declare -A opts
 
     opts["--target"]=local
+    opts["--no-deploy"]=false
+    opts["--no-helmify"]=false
 
     while [ $# -gt 0 ]; do
         case $1 in
@@ -80,6 +103,12 @@ parse_optional_params(){
                     opts["--target"]=$2
                     shift
                 fi 
+                ;;
+            --no-deploy)
+                opts["--no-deploy"]=true
+                ;;
+            --no-helmify)
+                opts["--no-helmify"]=true
                 ;;
         esac
         shift
@@ -155,25 +184,31 @@ deploy_unique(){
 
     echo "Deploying $NAME ..."
 
-    eval "$(parse_optional_params $@)"
-    
     case ${opts["--target"]} in
         # "EKS") deploy_eks $NAME;;
         # "Azure") deploy_azure $NAME;;
         # "GCP") deploy_gcp $NAME;;
         *)
-            rm -rf $CURRENT_DIR/$NAME/deploy
-            helmify $NAME
 
-            if [ $? -eq 1 ]; then
-                return 1
+            if [ "${opts["--no-helmify"]}" == false ]; then
+                rm -rf $CURRENT_DIR/$NAME/deploy
+                
+                helmify $NAME
+
+                if [ $? -eq 1 ]; then
+                    print_error "Error occured helmifying $NAME. Please try again or use 'kubefs --help' for more information."
+                    return 1
+                fi
             fi
 
-            helm upgrade --install $NAME $CURRENT_DIR/$NAME/deploy 
+            if [ "${opts["--no-deploy"]}" == false ]; then
 
-            if [ $? -eq 1 ]; then
-                print_error "Error occured deploying $NAME. Please try again or use 'kubefs --help' for more information."
-                return 1
+                helm upgrade --install $NAME $CURRENT_DIR/$NAME/deploy
+
+                if [ $? -eq 1 ]; then
+                    print_error "Error occured deploying $NAME. Please try again or use 'kubefs --help' for more information."
+                    return 1
+                fi
             fi
             ;;
     esac
